Struct tui_logger::TuiLoggerSmartWidget

pub struct TuiLoggerSmartWidget<'a> { /* private fields */ }

The Smart Widget combines the TuiLoggerWidget and the TuiLoggerTargetWidget into a nice combo, where the TuiLoggerTargetWidget can be shown/hidden.

In the title the number of logging messages/s in the whole buffer is shown.

Implementations

impl<'a> TuiLoggerSmartWidget<'a>

pub fn highlight_style(self, style: Style) -> Self

pub fn border_style(self, style: Style) -> Self

pub fn border_type(self, border_type: BorderType) -> Self

pub fn style(self, style: Style) -> Self

pub fn style_error(self, style: Style) -> Self

pub fn style_warn(self, style: Style) -> Self

pub fn style_info(self, style: Style) -> Self

pub fn style_trace(self, style: Style) -> Self

pub fn style_debug(self, style: Style) -> Self

pub fn style_off(self, style: Style) -> Self

pub fn style_hide(self, style: Style) -> Self

pub fn style_show(self, style: Style) -> Self

pub fn output_separator(self, sep: char) -> Self

Separator character between field. Default is ‘:’

pub fn output_timestamp(self, fmt: Option<String>) -> Self

The format string can be defined as described in https://docs.rs/chrono/0.4.19/chrono/format/strftime/index.html

If called with None, timestamp is not included in output.

Default is %H:%M:%S

pub fn output_level(self, level: Option<TuiLoggerLevelOutput>) -> Self

Possible values are

TuiLoggerLevelOutput::Long => DEBUG/TRACE/…

TuiLoggerLevelOutput::Abbreviated => D/T/…

If called with None, level is not included in output.

Default is Long

pub fn output_target(self, enabled: bool) -> Self

Enables output of target field of event

Default is true

pub fn output_file(self, enabled: bool) -> Self

Enables output of file field of event

Default is true

pub fn output_line(self, enabled: bool) -> Self

Enables output of line field of event

Default is true

pub fn title_target<T>(self, title: T) -> Self
where T: Into<Line<'a>>,

pub fn title_log<T>(self, title: T) -> Self
where T: Into<Line<'a>>,

pub fn state(self, state: &TuiWidgetState) -> Self